Re-merge Python wrapper
